DjangoProject/requirements.txt
Django~=4.2
python-dotenv~=1.0.0
djangorestframework~=3.14.0
django-filter~=23.2